SNPs v1.02
This version of SNPs should be able to take any legal SAM file as an
input. It was tested on both bowtie and bowtie2 output. In addition to
call of the 4 bases+N this version also includes information about
insertions and deletions, if information is available from the
alignment program.